% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{analyze_fragments}
\alias{analyze_fragments}
\title{Analyze Fragments}
\usage{
analyze_fragments(dat_aligned, fragment_size)
}
\arguments{
\item{dat_aligned}{A matrix of aligned sequences.}

\item{fragment_size}{The size of each fragment to analyze.}
}
\value{
A list containing percentage of alphabetic characters and a final table with fragment analysis.
}
\description{
This function analyzes fragments of the aligned data and returns statistics on the fragments.
}
