\name{clamp.crf}
\alias{clamp.crf}

\title{
Make clamped CRF data structure
}

\description{
Generate clamped CRF data structure by fixing the states of some nodes
}

\usage{
clamp.crf(crf, clamped)
}

\arguments{
  \item{crf}{The CRF data structure generated by \code{\link{make.crf}}}
  \item{clamped}{The vector of fixed states of nodes}
}

\details{
The function will generate a clamped CRF data structure from a given CRF data structure by fixing the states of some nodes.
The vector \code{clamped} contains the desired state for each node while zero means the state is not fixed.
The node and edge potentials are updated to the conditional potentials based on the clamped vector.
}

\value{
The function will return a CRF data structure with additional components:
\item{original}{The original CRF data.}
\item{clamped}{The vector of fixed states of nodes.}
\item{node.id}{The vector of the original node ids for nodes in the new CRF data.}
\item{node.map}{The vector of the new node ids for nodes in the original CRF data.}
\item{edge.id}{The vector of the original edge ids for edges in the new CRF data.}
\item{edge.map}{The vector of the new edge ids for edges in the original CRF data.}
}

\seealso{
\code{\link{make.crf}}, \code{\link{sub.crf}}
}

\examples{
library(CRF)
data(Small)
crf <- clamp.crf(small.crf, c(0, 0, 1, 1))

}

\keyword{ make.crf }
