% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{count_unique_characters}
\alias{count_unique_characters}
\title{Count Unique Characters}
\usage{
count_unique_characters(column)
}
\arguments{
\item{column}{A vector representing a column from a matrix.}
}
\value{
The count of unique characters excluding dashes.
}
\description{
This function counts the number of unique characters in a given column, excluding dashes ("-").
}
