\name{crf.nll}
\alias{crf.nll}
\title{Calculate CRF negative log likelihood}
\usage{
crf.nll(par, crf, instances, node.fea = NaN, edge.fea = NaN,
  node.ext = NaN, edge.ext = NaN, infer.method = infer.chain, ...)
}
\arguments{
  \item{crf}{}

  \item{par}{}

  \item{instances}{}

  \item{node.fea}{}

  \item{edge.fea}{}

  \item{node.ext}{}

  \item{edge.ext}{}

  \item{infer.method}{}

  \item{...}{}
}
\value{
This function will return the value of CRF negative
log-likelihood.
}
\description{
Calculate the negative log likelihood of CRF model
}
\details{
Calculate the negative log likelihood of CRF model
}

