\name{crf.update}
\alias{crf.update}
\title{Update CRF potentials}
\usage{
crf.update(crf, node.fea = NaN, edge.fea = NaN, node.ext = NaN,
  edge.ext = NaN)
}
\arguments{
  \item{crf}{}

  \item{node.fea}{}

  \item{edge.fea}{}

  \item{node.ext}{}

  \item{edge.ext}{}
}
\value{
This function will directly modify the CRF. Do not use the
returned value.
}
\description{
Update node.pot and edge.pot of CRF model
}
\details{
Update node.pot and edge.pot of CRF model
}

