% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_similar_rows}
\alias{find_similar_rows}
\title{Find Similar Rows}
\usage{
find_similar_rows(matrix_data)
}
\arguments{
\item{matrix_data}{A matrix of aligned sequences.}
}
\value{
A list of rows with similar values.
}
\description{
This function finds rows in a matrix that have similar values, ignoring dashes.
}
