\name{gradient}
\alias{gradient}
\title{Calculate CRF negative log-likelihood gradient}
\usage{
gradient(par, crf, ...)
}
\arguments{
  \item{par}{}

  \item{crf}{}

  \item{...}{}
}
\value{
This function will return the gradient of CRF negative
log-likelihood.
}
\description{
Calculate the gradient of negative log likelihood of CRF
model
}
\details{
Calculate the gradient of negative log likelihood of CRF
model. This function is used by optimization algorithm in
training.
}

