% Generated by roxygen2 (4.0.1): do not edit by hand
\name{infer.junction}
\alias{infer.junction}
\title{Inference method for low-treewidth graphs}
\usage{
infer.junction(crf)
}
\arguments{
\item{crf}{The CRF}
}
\value{
This function will return a list with components:
  \item{node.bel}{Node belief. It is a matrix with \code{crf$n.nodes} rows and \code{crf$max.state} columns.}
  \item{edge.bel}{Edge belief. It is a list of matrices. The size of list is \code{crf$n.edges} and
    the matrix \code{i} has \code{crf$n.states[crf$edges[i,1]]} rows and \code{crf$n.states[crf$edges[i,2]]} columns.}
  \item{logZ}{The logarithmic value of CRF normalization factor Z.}
}
\description{
Computing the partition function and marginal probabilities
}
\details{
Exact decoding for low-treewidth graphs using junction trees
}
\examples{
library(CRF)
data(Small)
i <- infer.junction(Small$crf)
}

