\name{make.par}
\alias{make.par}
\title{Make CRF parameters}
\usage{
make.par(crf, n.par = 1)
}
\arguments{
  \item{crf}{}

  \item{n.par}{}
}
\value{
This function will return the same CRF.
}
\description{
Make the data structure of parameters
}
\details{
Make the data structure of parameters need for modeling and
training
}

