\name{mrf.nll}
\alias{mrf.nll}
\title{Calculate MRF negative log-likelihood}
\usage{
mrf.nll(par, crf, instances, infer.method = infer.chain, ...)
}
\arguments{
  \item{crf}{}

  \item{par}{}

  \item{instances}{}

  \item{infer.method}{}

  \item{...}{}
}
\value{
This function will return the value of MRF negative
log-likilihood.
}
\description{
Calculate the negative log-likelihood of MRF model
}
\details{
Calculate the negative log-likelihood of MRF model
}

