\name{mrf.stat}
\alias{mrf.stat}
\title{Calculate MRF sufficient statistics}
\usage{
mrf.stat(crf, instances)
}
\arguments{
  \item{crf}{}

  \item{instances}{}
}
\value{
This function will return the value of MRF sufficient
statistics.
}
\description{
Calculate the sufficient statistics of MRF model
}
\details{
Calculate the sufficient statistics of MRF model
}

