\name{mrf.update}
\alias{mrf.update}
\title{Update MRF potentials}
\usage{
mrf.update(crf)
}
\arguments{
  \item{crf}{}
}
\value{
This function will directly modify the CRF. Do not use the
returned value.
}
\description{
Update node.pot and edge.pot of MRF model
}
\details{
Update node.pot and edge.pot of MRF model
}

