% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRF-package.R
\docType{data}
\name{Rain}
\alias{Rain}
\title{Rain data}
\format{A list containing two elements:
\itemize{
  \item \code{rain} A matrix of 28 columns containing raining data (1: rain, 2: sunny).
    Each row is an instance of 28 days for one month.
  \item \code{months} A vector containing the months of each instance.
}}
\usage{
data(Rain)
}
\description{
This data set gives an example of rain data used to train CRF and MRF models
}
\references{
Mark Schmidt. UGM: A Matlab toolbox for probabilistic undirected graphical models.
\url{http://www.cs.ubc.ca/~schmidtm/Software/UGM.html}, 2007.
}
\keyword{datasets}

