\name{sub.crf}
\alias{sub.crf}

\title{
Make sub CRF data structure
}

\description{
Generate sub CRF data structure by selecting some nodes
}

\usage{
sub.crf(crf, subset)
}

\arguments{
  \item{crf}{The CRF data structure generated by \code{\link{make.crf}}}
  \item{subset}{The vector of selected node ids}
}

\details{
The function will generate a CRF data structure from a given CRF data structure by selecting some nodes.
The vector \code{subset} contains the node ids selected to generate the new CRF data.
Unlike \code{\link{clamp.crf}}, the potentials of remainning nodes and edges are untouched.
}

\value{
The function will return a CRF data structure with additional components:
\item{original}{The original CRF data.}
\item{node.id}{The vector of the original node ids for nodes in the new CRF data.}
\item{node.map}{The vector of the new node ids for nodes in the original CRF data.}
\item{edge.id}{The vector of the original edge ids for edges in the new CRF data.}
\item{edge.map}{The vector of the new edge ids for edges in the original CRF data.}
}

\seealso{
\code{\link{make.crf}}, \code{\link{clamp.crf}}
}

\examples{
library(CRF)
data(Small)
crf <- sub.crf(small.crf, c(2, 3))

}

\keyword{ make.crf }
