\name{train.crf}
\alias{train.crf}
\title{Train CRF model}
\usage{
train.crf(crf, instances, node.fea = NaN, edge.fea = NaN, node.ext = NaN,
  edge.ext = NaN, trace = 0)
}
\arguments{
  \item{crf}{}

  \item{instances}{}

  \item{trace}{}

  \item{node.fea}{}

  \item{edge.fea}{}

  \item{node.ext}{}

  \item{edge.ext}{}
}
\value{
This function will return the same CRF.
}
\description{
Train the CRF model to estimate the parameters
}
\details{
This function train the CRF model.
}

