\name{train.mrf}
\alias{train.mrf}
\title{Train MRF model}
\usage{
train.mrf(crf, instances, trace = 0)
}
\arguments{
  \item{crf}{}

  \item{instances}{}

  \item{trace}{}
}
\value{
This function will return the same CRF.
}
\description{
Train the MRF model to estimate the parameters
}
\details{
This function trains the Markov Random Fields (MRF) model,
which is a simple variant of CRF model.
}

