\name{make.features}
\alias{make.features}
\title{Make CRF features}
\usage{
make.features(crf, n.nf = 1, n.ef = 1)
}
\arguments{
  \item{crf}{}

  \item{n.nf}{}

  \item{n.ef}{}
}
\value{
This function will return the same CRF.
}
\description{
Make the data structure of features
}
\details{
Make the data structure of features need for modeling and
training
}

