% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mrf.stat}
\alias{mrf.stat}
\title{Calculate MRF sufficient statistics}
\usage{
mrf.stat(crf, instances)
}
\arguments{
\item{crf}{The CRF}

\item{instances}{The training data matrix of MRF model}
}
\value{
This function will return the value of MRF sufficient statistics.
}
\description{
Calculate the sufficient statistics of MRF model
}
\details{
This function calculates the sufficient statistics of MRF model. This function
much be called before the first calling to \code{\link{mrf.nll}}.
In the training data matrix \code{instances}, each row is an instance and
each column corresponds a node in CRF.
}
\seealso{
\code{\link{mrf.nll}}, \code{\link{train.mrf}}
}

